2024.03.12

+++ Java/20240312/thread/src/main/java/com/nhnacademy/SharedCounter.java
@@ -2,10 +2,10 @@ package com.nhnacademy;
 
 public class SharedCounter extends Thread {
     private SharedCount sharedCount;
-    private int count;
-    private int maxCount;
+    private long count;
+    private long maxCount;
 
-    public SharedCounter(String name, int maxCount, SharedCount sharedCount) {
+    public SharedCounter(String name, long maxCount, SharedCount sharedCount) {
         setName(name);
         this.sharedCount = sharedCount;
         this.maxCount = maxCount;
